Create ArgusEyesV0.21.py
Discovered some of the sites have volatile HTML elements, I'm going to use RSS feeds from those websites to work-around the problem so this scraper remains usable over the development time. The ones with this work-around will be annotated within the FXN names (i.e., 'nyt_US_RSS(fo)'). 

those sites, so far are:
    +NYT - POLITICS
    +NYT - TECH
    +NYT - WORLD
    +BBC - UK

I have yet to discover a way to predict and 'fix' the code for these volatile elemets as they change. The work around will do for now as I build this part of the project. After completing the project of which this is a piece, I'll go back and try to find a way to fix this within ArgusEyes to use the main websites again.