Create ArgusEyesV0.24.py
Added new function for yahoo finance site and RSS feed. Still investigating how to streamline. May have to use a global, but I think I can make-do with a class. Idea is to write all of these to a list or lists and multithreading using 2 or 3 threads to improve overall performance. Then it will combine the lists in the end and write to a file.